frontend/requirements.txt
asyncpg==0.30.0
streamlit==1.44.1
requests==2.32.3
ollama==0.4.8
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pymilvus==2.5.6
pymilvus[model]
pydantic==2.9.0
